refactor: move dto/auth
